pyproject.toml
[tool.poetry]
name = "tele-notifier"
version = "0.1.0"
description = ""
authors = ["Thong Nguyen <thoongnv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
redis = "^5.0.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
